refactor(quant,models): AWQ byte-math to rmlx-quant; shared PARO/embedding-quant loaders
Move the seven AWQ pack/unpack byte-math fns (awq_unpack_word, mlx_pack_word,
convert_awq_qweight, convert_awq_qzeros_to_biases, f16_bits_to_f32,
f32_to_f16_bits, quantize_f16_affine_int4) verbatim out of
qwen3_5_moe/loader.rs into a new rmlx-quant::awq module. They are pure
byte-level conversions with no mlx/Array dependency, so rmlx-quant stays a
leaf — no new edges, no cycle.

Add shared load_util helpers — ParoParts + load_paro_parts and
quantize_embedding_int4 — and route both the qwen3_5_moe and gemma4 PARO /
int4-embedding load paths through them, deleting the per-arch copies.

Pure code motion: the moved fns are byte-identical (reviewer-confirmed, only
visibility/location differs), and temp=0 decode on both PARO snapshots
(Qwen3.6-27B-PARO, gemma-4-31B-it-PARO) is byte-identical pre/post.
Net ~544 LOC removed across the two loaders. make ci green.
