+++ crates/rmlx-models/src/qwen3_5_moe/tests.rs
@@ -15,9 +15,9 @@ use super::*;
 // and are needed by multiple test functions.
 use rmlx_mlx::{add, divide, softmax, sum_axis, Array, Device, Dtype};
 
-// pub(crate) items: loader helpers re-exported from mod.rs but not visible
-// via `use super::*` (glob only includes `pub` items).
-use super::{
+// AWQ/F16 byte-math now lives in `rmlx_quant::awq`. The MLX-integration tests
+// below still call them to build inputs for `quantized_matmul`/`embed_lookup`.
+use rmlx_quant::awq::{
     convert_awq_qweight, convert_awq_qzeros_to_biases, f16_bits_to_f32, f32_to_f16_bits,
     quantize_f16_affine_int4,
 };
@@ -322,11 +322,15 @@ fn prompt_cache_identical_prompt_is_exact_not_partial() {
     );
 }
 
-/// Validate AWQ → MLX weight conversion + quantized_matmul with a synthetic example.
+/// Validate AWQ → MLX weight conversion feeds a correct `quantized_matmul`.
 ///
 /// Synthetic: in=128, out=8, bits=4, group_size=128, num_groups=1.
 /// Weight values after dequant: nibble[i, o] = (i + o) % 8, scale=1.0, zero=0.
 /// So dequant(x=ones): out[o] = sum_{i=0}^{127} (i + o) % 8.
+///
+/// The pure byte-layout assertions for `convert_awq_qweight` /
+/// `convert_awq_qzeros_to_biases` live in `rmlx-quant` (`awq_tests.rs`); this
+/// test covers the MLX-integration path that `rmlx-quant` cannot (no mlx dep).
 ///
 /// MLX requires group_size ∈ {32, 64, 128} — smaller values have no kernel.
 #[test]
@@ -342,7 +346,7 @@ fn prompt_cache_identical_prompt_is_exact_not_partial() {
     clippy::unwrap_used,
     reason = "Mutex critical section is panic-free, so PoisonError is structurally unreachable; remaining Option/Result unwrap is on values established by construction earlier in this fn"
 )]
-fn paro_weight_conversion_roundtrip() {
+fn paro_weight_conversion_matmul() {
     use rmlx_mlx::{Array, Device, Dtype};
 
     let in_f = 128usize;
@@ -378,40 +382,8 @@ fn paro_weight_conversion_roundtrip() {
     // Convert AWQ qweight → MLX layout.
     let mlx_weight_bytes =
         convert_awq_qweight(&qweight_bytes, in_f, out_f, bits).expect("convert_awq_qweight");
-    // Expected shape: [out=8, in*bits/32=1] words.
-    assert_eq!(mlx_weight_bytes.len(), out_f * (in_f * bits / 32) * 4);
-
-    // Verify: unpack MLX weight and check nibbles match transpose [in, out] → [out, in].
-    let words_per_out = in_f * bits / 32; // = 128*4/32 = 16
-    #[allow(clippy::needless_range_loop)]
-    for o in 0..out_f {
-        for j in 0..words_per_out {
-            let off = (o * words_per_out + j) * 4;
-            let word = u32::from_le_bytes([
-                mlx_weight_bytes[off],
-                mlx_weight_bytes[off + 1],
-                mlx_weight_bytes[off + 2],
-                mlx_weight_bytes[off + 3],
-            ]);
-            // MLX sequential: nibble at position k = input element (j*8 + k).
-            for k in 0..8usize {
-                let in_idx = j * 8 + k;
-                if in_idx >= in_f {
-                    break;
-                }
-                let nibble = ((word >> (k * 4)) & 0xF) as u8;
-                let expected = nibble_matrix[in_idx][o];
-                assert_eq!(
-                    nibble, expected,
-                    "MLX weight nibble[out={o}, in={in_idx}]: expected {expected}, got {nibble}"
-                );
-            }
-        }
-    }
 
     // Build scales (F16 = 1.0) and zeros (0) for num_groups=1, out_f=8.
-    // AWQ scales shape: [num_groups=1, out_f=8] F16.
-    // AWQ qzeros shape: [num_groups=1, out_f*bits/32=1] I32 (zeros packed as nibbles).
     let scale_f16_bits: u16 = 0x3C00; // 1.0 in F16
     let mut scales_bytes = vec![0u8; num_groups * out_f * 2]; // [1, 8] F16
     for o in 0..out_f {
@@ -424,19 +396,6 @@ fn paro_weight_conversion_roundtrip() {
     let (scales_t_bytes, biases_t_bytes) =
         convert_awq_qzeros_to_biases(&qzeros_bytes, &scales_bytes, num_groups, out_f, bits)
             .expect("convert_awq_qzeros_to_biases");
-
-    // scales_t: [out=8, num_groups=1] F16 = all 1.0
-    // biases_t: [out=8, num_groups=1] F16 = -1.0 * 0 = 0.0
-    for o in 0..out_f {
-        let s_bits = u16::from_le_bytes([scales_t_bytes[o * 2], scales_t_bytes[o * 2 + 1]]);
-        assert_eq!(s_bits, scale_f16_bits, "scale[o={o}] must be 1.0 F16");
-        let b_bits = u16::from_le_bytes([biases_t_bytes[o * 2], biases_t_bytes[o * 2 + 1]]);
-        // -1.0 * 0 = -0.0 (F16 0x8000) or +0.0 (0x0000): both are zero.
-        assert!(
-            b_bits == 0u16 || b_bits == 0x8000u16,
-            "bias[o={o}] must be ±0.0 F16, got {b_bits:#06X}"
-        );
-    }
 
     // Call quantized_matmul with x=ones: expect out[o] = sum(nibble[i][o]) for i=0..7.
     // For our nibble matrix: out[o] = sum((i+o)%8, i=0..7) = 0+1+..+7 = 28 for all o.
@@ -505,93 +464,6 @@ fn paro_weight_conversion_roundtrip() {
             (v - expected).abs() < tol,
             "out[{o}]: expected {expected:.1}, got {v:.4} (diff={:.4})",
             (v - expected).abs()
-        );
-    }
-}
-
-/// Verify that `quantize_f16_affine_int4` matches MLX `mx.quantize` output.
-///
-/// Uses a known row of embed_tokens (row 760, first 128 elements) and checks
-/// that our Rust quantization produces the same scale/bias as Python.
-///
-/// Python reference (from paro_embed_check.py):
-/// row760[:128] scale ≈ -0.01029, bias ≈ 0.12354
-/// dequant[:8] = [0.02061, 3.05e-5, 3.05e-5, 3.05e-5, 3.05e-5, 3.05e-5, 0.01032, -0.01026]
-#[test]
-#[allow(
-    clippy::expect_used,
-    reason = "structural invariant: value present by construction in calling context; .expect() message documents the invariant"
-)]
-#[allow(
-    clippy::indexing_slicing,
-    reason = "bounds established by construction: buffer sized at init, loop indices bounded by slice length, or layer index validated before call"
-)]
-#[allow(
-    clippy::unwrap_used,
-    reason = "Mutex critical section is panic-free, so PoisonError is structurally unreachable; remaining Option/Result unwrap is on values established by construction earlier in this fn"
-)]
-fn quantize_f16_affine_int4_matches_python() {
-    // Synthesize a known group: row with min≈-0.034, max≈0.124.
-    // These bounds are from the first 128 elements of embed_tokens.weight[760].
-    // We don't need the actual model file — we can test the formula directly.
-    let group_size = 128usize;
-    let n = group_size;
-
-    // Build a synthetic F16 row with known min=-0.08 and max=0.08.
-    // scale = -(0.08 - (-0.08)) / 15 = -0.16/15 ≈ -0.010667
-    // bias = max = 0.08
-    let min_val = -0.08_f32;
-    let max_val = 0.08_f32;
-    let mut row_f32: Vec<f32> = (0..n)
-        .map(|i| min_val + (max_val - min_val) * (i as f32) / ((n - 1) as f32))
-        .collect();
-
-    // Encode as F16 bytes.
-    let mut row_bytes = vec![0u8; n * 2];
-    for (i, &v) in row_f32.iter().enumerate() {
-        let bits = f32_to_f16_bits(v);
-        row_bytes[i * 2..i * 2 + 2].copy_from_slice(&bits.to_le_bytes());
-    }
-    // Re-read the actual F16 values (may differ slightly from f32 due to F16 precision).
-    for i in 0..n {
-        let bits = u16::from_le_bytes([row_bytes[i * 2], row_bytes[i * 2 + 1]]);
-        row_f32[i] = f16_bits_to_f32(bits);
-    }
-    let actual_min = row_f32.iter().cloned().fold(f32::INFINITY, f32::min);
-    let actual_max = row_f32.iter().cloned().fold(f32::NEG_INFINITY, f32::max);
-
-    let (wq_bytes, sc_bytes, bi_bytes) =
-        quantize_f16_affine_int4(&row_bytes, 1, n, group_size).expect("quantize");
-
-    let scale_bits = u16::from_le_bytes([sc_bytes[0], sc_bytes[1]]);
-    let bias_bits = u16::from_le_bytes([bi_bytes[0], bi_bytes[1]]);
-    let scale_f32 = f16_bits_to_f32(scale_bits);
-    let bias_f32 = f16_bits_to_f32(bias_bits);
-
-    let expected_scale = -(actual_max - actual_min) / 15.0;
-    let expected_bias = actual_max;
-    let tol = 0.001_f32;
-    assert!(
-        (scale_f32 - f16_bits_to_f32(f32_to_f16_bits(expected_scale))).abs() < tol,
-        "scale mismatch: got={scale_f32:.6}, expected≈{expected_scale:.6}"
-    );
-    assert!(
-        (bias_f32 - f16_bits_to_f32(f32_to_f16_bits(expected_bias))).abs() < tol,
-        "bias mismatch: got={bias_f32:.6}, expected≈{expected_bias:.6}"
-    );
-
-    // Dequant and check round-trip error is within 1 quantization step.
-    let step = (actual_max - actual_min) / 15.0;
-    for (i, &orig) in row_f32.iter().enumerate().take(n) {
-        let word_idx = i / 8;
-        let nibble_pos = i % 8;
-        let word = u32::from_le_bytes(wq_bytes[word_idx * 4..word_idx * 4 + 4].try_into().unwrap());
-        let nibble = (word >> (nibble_pos * 4)) & 0xF;
-        let dequant = (nibble as f32).mul_add(scale_f32, bias_f32);
-        let err = (dequant - orig).abs();
-        assert!(
-            err <= step + 0.001,
-            "dequant error at [{i}]: original={orig:.6}, dequant={dequant:.6}, err={err:.6}, step={step:.6}",
         );
     }
 }
@@ -1671,7 +1543,7 @@ fn paro_rotation_kernel_vs_python() {
     assert_eq!(half_hidden * 2, hidden, "theta half_hidden mismatch");
     let group_size = 128usize; // PARO Qwen3.6-27B always uses 128.
 
-    // Pre-compute cos/sin from F16 theta bytes (mirrors load_paro_linear).
+    // Pre-compute cos/sin from F16 theta bytes (mirrors load_paro_parts).
     let n_theta = krot * half_hidden;
     let mut cos_bytes = vec![0u8; n_theta * 2];
     let mut sin_bytes = vec![0u8; n_theta * 2];
